Added team spawn check
Now locates and put player at spawn of where he needs to be.
(Suggestion: We should use team.SetSpawn instead so we can set the spawn on the map instead. The only problem is that the fgd would only be able to support a limit so team limits might need to exist. No limit on this code but could easily implement.)